requirements.txt
requests
tulip-python>=5.7.2
python-dotenv
